+++ apps/client/src/views/admin/ListLogs.vue
@@ -1,10 +1,10 @@
 <script lang="ts" setup>
 import { computed, onMounted, ref } from 'vue'
 import type { Log } from '@cpn-console/shared'
+// @ts-ignore 'vue3-json-viewer' missing types
 import { JsonViewer } from 'vue3-json-viewer'
 import { useLogStore } from '@/stores/log.js'
 import { useSnackbarStore } from '@/stores/snackbar.js'
-// @ts-ignore 'vue3-json-viewer' missing types
 
 const adminLogStore = useLogStore()
 const snackbarStore = useSnackbarStore()
