--- apps/server/src/utils/const.ts
@@ -1,1 +1,0 @@
-export const tokenHeaderName = 'x-dso-token'
