+++ apps/server/src/utils/controller.ts
@@ -1,8 +1,7 @@
 import type { Cluster, Prisma, Project, ProjectMembers, ProjectRole } from '@prisma/client'
 import type { XOR } from '@cpn-console/shared'
-import { PROJECT_PERMS as PP, PROJECT_PERMS, projectIsLockedInfo } from '@cpn-console/shared'
+import { PROJECT_PERMS as PP, PROJECT_PERMS, projectIsLockedInfo, tokenHeaderName } from '@cpn-console/shared'
 import type { FastifyRequest } from 'fastify'
-import { tokenHeaderName } from './const.js'
 import type { UserDetails } from '@/types/index.js'
 import prisma from '@/prisma.js'
 import { logAdminToken, logUser } from '@/resources/user/business.js'
