+++ packages/shared/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@cpn-console/shared",
   "type": "module",
-  "version": "1.1.3",
+  "version": "1.1.4",
   "private": false,
   "description": "",
   "main": "dist/index.js",
