+++ packages/shared/src/contracts/cluster.ts
@@ -2,16 +2,18 @@ import type { ClientInferResponseBody } from '@ts-rest/core'
 import { z } from 'zod'
 import { apiPrefix, contractInstance } from '../api-client.js'
 import {
-  CreateClusterSchema,
-  DeleteClusterSchema,
-  GetClusterAssociatedEnvironmentsSchema,
-  GetClusterDetailsSchema,
-  GetClustersSchema,
-  UpdateClusterSchema,
+  CleanedClusterSchema,
+  ClusterDetailsSchema,
+  EnvironmentSchema,
+  OrganizationSchema,
+  UserSchema,
 } from '../schemas/index.js'
-import {
-  CoerceBooleanSchema,
-} from '../utils/schemas.js'
+import { CoerceBooleanSchema } from '../schemas/_utils.js'
+import { ErrorSchema, baseHeaders } from './_utils.js'
+
+export const ClusterParams = z.object({
+  clusterId: CleanedClusterSchema.shape.id,
+})
 
 export const clusterContract = contractInstance.router({
   listClusters: {
@@ -19,7 +21,11 @@ export const clusterContract = contractInstance.router({
     path: `${apiPrefix}/clusters`,
     summary: 'Get clusters',
     description: 'Retrieve clusters authorized for user',
-    responses: GetClustersSchema.responses,
+    responses: {
+      200: z.array(CleanedClusterSchema),
+      401: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   createCluster: {
@@ -28,8 +34,13 @@ export const clusterContract = contractInstance.router({
     contentType: 'application/json',
     summary: 'Create cluster',
     description: 'Create new cluster.',
-    body: CreateClusterSchema.body,
-    responses: CreateClusterSchema.responses,
+    body: ClusterDetailsSchema.omit({ id: true }),
+    responses: {
+      201: ClusterDetailsSchema,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   getClusterDetails: {
@@ -37,8 +48,13 @@ export const clusterContract = contractInstance.router({
     path: `${apiPrefix}/clusters/:clusterId`,
     summary: 'Get cluster details',
     description: 'Retrieved details of a cluster.',
-    pathParams: GetClusterDetailsSchema.params,
-    responses: GetClusterDetailsSchema.responses,
+    pathParams: ClusterParams,
+    responses: {
+      200: ClusterDetailsSchema,
+      401: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   getClusterEnvironments: {
@@ -46,8 +62,19 @@ export const clusterContract = contractInstance.router({
     path: `${apiPrefix}/clusters/:clusterId/environments`,
     summary: 'Get cluster envs',
     description: 'Retrieved environments linked to a cluster.',
-    pathParams: GetClusterAssociatedEnvironmentsSchema.params,
-    responses: GetClusterAssociatedEnvironmentsSchema.responses,
+    pathParams: ClusterParams,
+    responses: {
+      200: z.array(z.object({
+        organization: OrganizationSchema.shape.name,
+        // TODO: Remettre `ProjectSchemaV2.shape.name` mais attention aux projets non compatibles
+        project: z.string(),
+        name: EnvironmentSchema.shape.name,
+        owner: UserSchema.shape.email.optional(),
+      })),
+      401: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   updateCluster: {
@@ -55,9 +82,15 @@ export const clusterContract = contractInstance.router({
     path: `${apiPrefix}/clusters/:clusterId`,
     summary: 'Update cluster',
     description: 'Update a cluster by its ID.',
-    pathParams: UpdateClusterSchema.params,
-    body: UpdateClusterSchema.body,
-    responses: UpdateClusterSchema.responses,
+    pathParams: ClusterParams,
+    body: ClusterDetailsSchema.omit({ id: true, label: true }).partial(),
+    responses: {
+      200: ClusterDetailsSchema,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   deleteCluster: {
@@ -66,10 +99,21 @@ export const clusterContract = contractInstance.router({
     summary: 'Delete cluster',
     description: 'Delete a cluster by its ID.',
     query: z.object({ force: CoerceBooleanSchema.optional() }),
-    pathParams: DeleteClusterSchema.params,
+    pathParams: ClusterParams,
     body: null,
-    responses: DeleteClusterSchema.responses,
+    responses: {
+      204: z.string()
+        .nullable(),
+      400: ErrorSchema,
+      401: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
+}, {
+  baseHeaders,
 })
 
 export type ClusterAssociatedEnvironments = ClientInferResponseBody<typeof clusterContract.getClusterEnvironments, 200>
+export type CreateClusterBody = Zod.infer<typeof clusterContract.createCluster.body>
+export type UpdateClusterBody = Zod.infer<typeof clusterContract.updateCluster.body>
