+++ packages/shared/src/contracts/environment.ts
@@ -1,11 +1,10 @@
 import type { ClientInferRequest } from '@ts-rest/core'
+import { z } from 'zod'
 import { apiPrefix, contractInstance } from '../api-client.js'
 import {
-  CreateEnvironmentSchema,
-  DeleteEnvironmentSchema,
-  GetEnvironmentsSchema,
-  UpdateEnvironmentSchema,
+  EnvironmentSchema,
 } from '../schemas/index.js'
+import { ErrorSchema, baseHeaders } from './_utils.js'
 
 export const environmentContract = contractInstance.router({
   createEnvironment: {
@@ -14,8 +13,13 @@ export const environmentContract = contractInstance.router({
     contentType: 'application/json',
     summary: 'Create environment',
     description: 'Create new environment.',
-    body: CreateEnvironmentSchema.body,
-    responses: CreateEnvironmentSchema.responses,
+    body: EnvironmentSchema.omit({ id: true }),
+    responses: {
+      201: EnvironmentSchema,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   listEnvironments: {
@@ -23,8 +27,18 @@ export const environmentContract = contractInstance.router({
     path: `${apiPrefix}/environments`,
     summary: 'Get environments',
     description: 'Retrieved project environments.',
-    query: GetEnvironmentsSchema.query,
-    responses: GetEnvironmentsSchema.responses,
+    query: z.object({
+      projectId: z.string()
+        .uuid(),
+    }),
+    responses: {
+      200: EnvironmentSchema.array(),
+      400: ErrorSchema,
+      401: ErrorSchema,
+      403: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   updateEnvironment: {
@@ -32,9 +46,19 @@ export const environmentContract = contractInstance.router({
     path: `${apiPrefix}/environments/:environmentId`,
     summary: 'Update environment',
     description: 'Update a environment by its ID.',
-    pathParams: UpdateEnvironmentSchema.params,
-    body: UpdateEnvironmentSchema.body,
-    responses: UpdateEnvironmentSchema.responses,
+    pathParams: z.object({
+      environmentId: z.string()
+        .uuid(),
+    }),
+    body: EnvironmentSchema.pick({ quotaId: true }),
+    responses: {
+      200: EnvironmentSchema,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      403: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   deleteEnvironment: {
@@ -42,12 +66,23 @@ export const environmentContract = contractInstance.router({
     path: `${apiPrefix}/environments/:environmentId`,
     summary: 'Delete environment',
     description: 'Delete a environment by its ID.',
-    pathParams: DeleteEnvironmentSchema.params,
     body: null,
-    responses: DeleteEnvironmentSchema.responses,
+    pathParams: z.object({
+      environmentId: z.string()
+        .uuid(),
+    }),
+    responses: {
+      204: null,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      403: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
+}, {
+  baseHeaders,
 })
 
 export type CreateEnvironmentBody = ClientInferRequest<typeof environmentContract.createEnvironment>['body']
-
 export type UpdateEnvironmentBody = ClientInferRequest<typeof environmentContract.updateEnvironment>['body']
