+++ packages/shared/src/contracts/project-member.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 import { MemberSchema, apiPrefix, contractInstance } from '../index.js'
-import { ErrorSchema } from '../schemas/utils.js'
+import { ErrorSchema, baseHeaders } from './_utils.js'
 
 export const projectMemberContract = contractInstance.router({
   listMembers: {
@@ -8,7 +8,7 @@ export const projectMemberContract = contractInstance.router({
     path: `${apiPrefix}/projects/:projectId/members`,
     pathParams: z.object({ projectId: z.string().uuid() }),
     responses: {
-      200: z.lazy(() => MemberSchema.array()),
+      200: MemberSchema.array(),
       400: ErrorSchema,
       401: ErrorSchema,
       403: ErrorSchema,
@@ -21,7 +21,7 @@ export const projectMemberContract = contractInstance.router({
     body: z.object({ email: z.string() }).or(z.object({ userId: z.string() })),
     pathParams: z.object({ projectId: z.string().uuid() }),
     responses: {
-      201: z.lazy(() => MemberSchema.array()),
+      201: MemberSchema.array(),
       400: ErrorSchema,
       401: ErrorSchema,
       403: ErrorSchema,
@@ -37,7 +37,7 @@ export const projectMemberContract = contractInstance.router({
     }).array(),
     pathParams: z.object({ projectId: z.string().uuid() }),
     responses: {
-      200: z.lazy(() => MemberSchema.array()),
+      200: MemberSchema.array(),
       400: ErrorSchema,
       401: ErrorSchema,
       403: ErrorSchema,
@@ -53,11 +53,13 @@ export const projectMemberContract = contractInstance.router({
     }),
     body: null,
     responses: {
-      204: z.lazy(() => MemberSchema.array()),
+      204: MemberSchema.array(),
       400: ErrorSchema,
       401: ErrorSchema,
       403: ErrorSchema,
       404: ErrorSchema,
     },
   },
+}, {
+  baseHeaders,
 })
