+++ packages/shared/src/contracts/project-role.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 import { RoleSchema, apiPrefix, contractInstance } from '../index.js'
-import { ErrorSchema } from '../schemas/utils.js'
+import { ErrorSchema, baseHeaders } from './_utils.js'
 
 export const projectRoleContract = contractInstance.router({
   listProjectRoles: {
@@ -8,7 +8,7 @@ export const projectRoleContract = contractInstance.router({
     path: `${apiPrefix}/projects/:projectId/roles`,
     pathParams: z.object({ projectId: z.string().uuid() }),
     responses: {
-      200: z.lazy(() => RoleSchema.array()),
+      200: RoleSchema.array(),
       400: ErrorSchema,
       401: ErrorSchema,
       403: ErrorSchema,
@@ -18,11 +18,11 @@ export const projectRoleContract = contractInstance.router({
   createProjectRole: {
     method: 'POST',
     path: `${apiPrefix}/projects/:projectId/roles`,
-    body: z.lazy(() => RoleSchema.omit({ position: true, id: true })),
+    body: RoleSchema.omit({ position: true, id: true }),
     pathParams: z.object({ projectId: z.string().uuid() }),
     responses: {
       // 200: z.any(),
-      200: z.lazy(() => RoleSchema),
+      200: RoleSchema,
       400: ErrorSchema,
       401: ErrorSchema,
       403: ErrorSchema,
@@ -34,10 +34,10 @@ export const projectRoleContract = contractInstance.router({
     path: `${apiPrefix}/projects/:projectId/roles`,
     pathParams: z.object({ projectId: z.string().uuid() }),
     // body: z.any(),
-    body: z.lazy(() => RoleSchema.partial({ name: true, permissions: true, position: true }).array()),
+    body: RoleSchema.partial({ name: true, permissions: true, position: true }).array(),
     responses: {
       // 200: z.any(),
-      200: z.lazy(() => RoleSchema.array()),
+      200: RoleSchema.array(),
       400: ErrorSchema,
       401: ErrorSchema,
       403: ErrorSchema,
@@ -72,4 +72,6 @@ export const projectRoleContract = contractInstance.router({
       404: ErrorSchema,
     },
   },
+}, {
+  baseHeaders,
 })
