+++ packages/shared/src/contracts/project-service.ts
@@ -1,8 +1,7 @@
+import { z } from 'zod'
 import { apiPrefix, contractInstance } from '../api-client.js'
-import {
-  GetProjectServicesSchema,
-  UpdateProjectServicesSchema,
-} from '../schemas/index.js'
+import { ProjectParams, ServiceSchema, permissionTarget, pluginUpdateBody } from '../index.js'
+import { ErrorSchema, baseHeaders } from './_utils.js'
 
 export const projectServiceContract = contractInstance.router({
   getServices: {
@@ -10,9 +9,14 @@ export const projectServiceContract = contractInstance.router({
     path: `${apiPrefix}/project/:projectId/services`,
     summary: 'Get Project\'s services',
     description: 'Get all informations about services related to a project.',
-    query: GetProjectServicesSchema.query,
-    pathParams: GetProjectServicesSchema.params,
-    responses: GetProjectServicesSchema.responses,
+    query: z.object({ permissionTarget }),
+    pathParams: ProjectParams,
+    responses: {
+      200: ServiceSchema.array(),
+      400: ErrorSchema,
+      401: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   updateProjectServices: {
@@ -20,8 +24,15 @@ export const projectServiceContract = contractInstance.router({
     path: `${apiPrefix}/project/:projectId/services`,
     summary: 'Update project service configuration',
     description: 'Update project service configuration',
-    pathParams: UpdateProjectServicesSchema.params,
-    body: UpdateProjectServicesSchema.body,
-    responses: UpdateProjectServicesSchema.responses,
+    pathParams: ProjectParams,
+    body: pluginUpdateBody,
+    responses: {
+      204: null,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
+}, {
+  baseHeaders,
 })
