+++ packages/shared/src/contracts/quota.ts
@@ -1,12 +1,8 @@
 import type { ClientInferRequest, ClientInferResponseBody } from '@ts-rest/core'
+import { z } from 'zod'
 import { apiPrefix, contractInstance } from '../api-client.js'
-import {
-  CreateQuotaSchema,
-  DeleteQuotaSchema,
-  ListQuotaEnvironmentsSchema,
-  ListQuotasSchema,
-  UpdateQuotaSchema,
-} from '../schemas/index.js'
+import { QuotaSchema } from '../schemas/index.js'
+import { ErrorSchema, baseHeaders } from './_utils.js'
 
 export const quotaContract = contractInstance.router({
   listQuotas: {
@@ -14,7 +10,10 @@ export const quotaContract = contractInstance.router({
     path: `${apiPrefix}/quotas`,
     summary: 'Get quotas',
     description: 'Retrieve all quotas.',
-    responses: ListQuotasSchema.responses,
+    responses: {
+      200: z.array(QuotaSchema),
+      500: ErrorSchema,
+    },
   },
 
   createQuota: {
@@ -23,17 +22,37 @@ export const quotaContract = contractInstance.router({
     contentType: 'application/json',
     summary: 'Create quota',
     description: 'Create new quota.',
-    body: CreateQuotaSchema.body,
-    responses: CreateQuotaSchema.responses,
+    body: QuotaSchema.omit({ id: true }).partial({ stageIds: true }),
+    responses: {
+      201: QuotaSchema,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      403: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   listQuotaEnvironments: {
     method: 'GET',
     path: `${apiPrefix}/quotas/:quotaId/environments`,
-    pathParams: ListQuotaEnvironmentsSchema.params,
     summary: 'Get a quota\'s environment',
     description: 'Retrieved environments associated to a quota.',
-    responses: ListQuotaEnvironmentsSchema.responses,
+    pathParams: z.object({
+      quotaId: z.string()
+        .uuid(),
+    }),
+    responses: {
+      200: z.array(z.object({
+        organization: z.string(),
+        project: z.string(),
+        name: z.string(),
+        stage: z.string(),
+        owner: z.string(),
+      })),
+      403: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   updateQuota: {
@@ -41,9 +60,22 @@ export const quotaContract = contractInstance.router({
     path: `${apiPrefix}/quotas/:quotaId`,
     summary: 'Update a quota privacy',
     description: 'Update a quota privacy.',
-    pathParams: UpdateQuotaSchema.params,
-    body: UpdateQuotaSchema.body,
-    responses: UpdateQuotaSchema.responses,
+    pathParams: z.object({
+      quotaId: z.string()
+        .uuid(),
+    }),
+    body: QuotaSchema.pick({
+      isPrivate: true,
+      cpu: true,
+      memory: true,
+      stageIds: true,
+      name: true,
+    }).partial(),
+    responses: {
+      200: QuotaSchema,
+      403: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   deleteQuota: {
@@ -51,10 +83,19 @@ export const quotaContract = contractInstance.router({
     path: `${apiPrefix}/quotas/:quotaId`,
     summary: 'Delete quota',
     description: 'Delete a quota by its ID.',
-    pathParams: DeleteQuotaSchema.params,
     body: null,
-    responses: DeleteQuotaSchema.responses,
+    pathParams: z.object({
+      quotaId: z.string()
+        .uuid(),
+    }),
+    responses: {
+      204: null,
+      403: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
+}, {
+  baseHeaders,
 })
 
 export type CreateQuotaBody = ClientInferRequest<typeof quotaContract.createQuota>['body']
