+++ packages/shared/src/contracts/repository.ts
@@ -1,12 +1,8 @@
 import type { ClientInferRequest } from '@ts-rest/core'
+import { z } from 'zod'
 import { apiPrefix, contractInstance } from '../api-client.js'
-import {
-  CreateRepoSchema,
-  DeleteRepoSchema,
-  GetReposSchema,
-  SyncRepoSchema,
-  UpdateRepoSchema,
-} from '../schemas/index.js'
+import { RepoSchema } from '../index.js'
+import { ErrorSchema, baseHeaders } from './_utils.js'
 
 export const repositoryContract = contractInstance.router({
   createRepository: {
@@ -15,27 +11,51 @@ export const repositoryContract = contractInstance.router({
     contentType: 'application/json',
     summary: 'Create repo',
     description: 'Create new repo.',
-    body: CreateRepoSchema.body,
-    responses: CreateRepoSchema.responses,
+    body: RepoSchema.omit({ id: true }),
+    responses: {
+      201: RepoSchema,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      403: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   listRepositories: {
     method: 'GET',
     path: `${apiPrefix}/repositories`,
-    query: GetReposSchema.query,
     summary: 'Get repos',
     description: 'Retrieved all repos.',
-    responses: GetReposSchema.responses,
+    query: z.object({
+      projectId: z.string()
+        .uuid(),
+    }),
+    responses: {
+      200: z.array(RepoSchema),
+      500: ErrorSchema,
+    },
   },
 
   syncRepository: {
     method: 'POST',
     path: `${apiPrefix}/repositories/:repositoryId/sync`,
-    body: SyncRepoSchema.body,
-    pathParams: SyncRepoSchema.params,
     summary: 'application/json',
     description: 'Trigger a gitlab synchronization for a repository',
-    responses: SyncRepoSchema.responses,
+    pathParams: z.object({
+      repositoryId: z.string()
+        .uuid(),
+    }),
+    body: z.object({
+      syncAllBranches: z.boolean(),
+      branchName: z.string().optional(),
+    }),
+    responses: {
+      204: null,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      403: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   updateRepository: {
@@ -43,9 +63,15 @@ export const repositoryContract = contractInstance.router({
     path: `${apiPrefix}/repositories/:repositoryId`,
     summary: 'Update repo',
     description: 'Update a repo by its ID.',
-    pathParams: UpdateRepoSchema.params,
-    body: UpdateRepoSchema.body,
-    responses: UpdateRepoSchema.responses,
+    pathParams: z.object({
+      repositoryId: z.string()
+        .uuid(),
+    }),
+    body: RepoSchema.partial(),
+    responses: {
+      200: RepoSchema,
+      500: ErrorSchema,
+    },
   },
 
   deleteRepository: {
@@ -54,9 +80,17 @@ export const repositoryContract = contractInstance.router({
     summary: 'Delete repo',
     description: 'Delete a repo by its ID.',
     body: null,
-    pathParams: DeleteRepoSchema.params,
-    responses: DeleteRepoSchema.responses,
+    pathParams: z.object({
+      repositoryId: z.string()
+        .uuid(),
+    }),
+    responses: {
+      204: null,
+      500: ErrorSchema,
+    },
   },
+}, {
+  baseHeaders,
 })
 
 export type CreateRepositoryBody = ClientInferRequest<typeof repositoryContract.createRepository>['body']
