+++ packages/shared/src/contracts/system.ts
@@ -1,13 +1,12 @@
 import type { ClientInferRequest, ClientInferResponseBody } from '@ts-rest/core'
+import { z } from 'zod'
 import { apiPrefix, contractInstance } from '../api-client.js'
 import {
-  GetHealthzSchema,
-  GetSystemPluginSchema,
-  GetVersionSchema,
-  ListSystemSettingsSchema,
-  UpdateSystemPluginSchema,
-  UpsertSystemSettingsSchema,
+  SystemSettingSchema,
+  pluginSchema,
+  pluginUpdateBody,
 } from '../schemas/index.js'
+import { ErrorSchema, baseHeaders } from './_utils.js'
 
 export const systemContract = contractInstance.router({
   getVersion: {
@@ -16,8 +15,10 @@ export const systemContract = contractInstance.router({
     summary: 'Get version',
     description: 'Retrieve api version.',
     responses: {
-      200: GetVersionSchema.responses['200'],
-      500: GetVersionSchema.responses['500'],
+      200: z.object({
+        version: z.string(),
+      }),
+      500: ErrorSchema,
     },
   },
 
@@ -27,8 +28,10 @@ export const systemContract = contractInstance.router({
     summary: 'Get health',
     description: 'Retrieve api health infos.',
     responses: {
-      200: GetHealthzSchema.responses['200'],
-      500: GetHealthzSchema.responses['500'],
+      200: z.object({
+        status: z.enum(['OK', 'KO']),
+      }),
+      500: ErrorSchema,
     },
   },
 })
@@ -39,7 +42,12 @@ export const systemPluginContract = contractInstance.router({
     path: `${apiPrefix}/system/plugins`,
     summary: 'Get plugins configuration',
     description: 'Get plugins configuration',
-    responses: GetSystemPluginSchema.responses,
+    responses: {
+      200: pluginSchema.array(),
+      400: ErrorSchema,
+      401: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   updatePluginsConfig: {
@@ -47,20 +55,31 @@ export const systemPluginContract = contractInstance.router({
     path: `${apiPrefix}/system/plugins`,
     summary: 'Update project service configuration',
     description: 'Update project service configuration',
-    body: UpdateSystemPluginSchema.body,
-    responses: UpdateSystemPluginSchema.responses,
+    body: pluginUpdateBody,
+    responses: {
+      204: null,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      403: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
-
+}, {
+  baseHeaders,
 })
 
 export const systemSettingsContract = contractInstance.router({
   listSystemSettings: {
     method: 'GET',
     path: `${apiPrefix}/system/settings`,
-    query: ListSystemSettingsSchema.query,
     summary: 'Get System Settings state',
     description: 'Get System Settings state',
-    responses: ListSystemSettingsSchema.responses,
+    query: SystemSettingSchema.pick({ key: true })
+      .partial(),
+    responses: {
+      200: SystemSettingSchema.array(),
+      500: ErrorSchema,
+    },
   },
 
   upsertSystemSetting: {
@@ -69,9 +88,16 @@ export const systemSettingsContract = contractInstance.router({
     contentType: 'application/json',
     summary: 'Update System Settings state',
     description: 'Update System Settings state',
-    body: UpsertSystemSettingsSchema.body,
-    responses: UpsertSystemSettingsSchema.responses,
+    body: SystemSettingSchema,
+    responses: {
+      201: SystemSettingSchema,
+      400: ErrorSchema,
+      401: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
+}, {
+  baseHeaders,
 })
 
 export type SystemSettings = ClientInferResponseBody<typeof systemSettingsContract.listSystemSettings, 200>
