+++ packages/shared/src/contracts/user.ts
@@ -2,23 +2,27 @@ import type { ClientInferRequest, ClientInferResponseBody } from '@ts-rest/core'
 import { z } from 'zod'
 import { apiPrefix, contractInstance } from '../api-client.js'
 import {
-  GetMatchingUsersSchema,
-  UserQuerySchema,
+  RoleNameCsvSchema,
   UserSchema,
 } from '../schemas/index.js'
-import { ErrorSchema } from '../schemas/utils.js'
+import { UuidOrCsvUuidSchema } from '../schemas/_utils.js'
+import { ErrorSchema, baseHeaders } from './_utils.js'
 
 export const userContract = contractInstance.router({
   getMatchingUsers: {
     method: 'GET',
     path: `${apiPrefix}/users/matching`,
-    query: GetMatchingUsersSchema.query,
+    query: z.object({
+      letters: z.string(),
+      notInProjectId: z.string().uuid().optional(),
+    }),
     summary: 'Get users by letters matching',
     description: 'Retrieved users by letters matching.',
     responses: {
-      200: z.lazy(() => UserSchema.array()),
+      200: UserSchema.array(),
       400: ErrorSchema,
       403: ErrorSchema,
+      500: ErrorSchema,
     },
   },
 
@@ -39,9 +43,18 @@ export const userContract = contractInstance.router({
     path: `${apiPrefix}/users`,
     summary: 'Get all users',
     description: 'Get all users.',
-    query: UserQuerySchema,
+    query: z.object({
+      adminRoles: RoleNameCsvSchema
+        .optional(),
+      adminRoleIds: UuidOrCsvUuidSchema
+        .optional(),
+      memberOfIds: UuidOrCsvUuidSchema
+        .optional(),
+      relationType: z.enum(['OR', 'AND'])
+        .optional(),
+    }),
     responses: {
-      200: z.lazy(() => UserSchema.array()),
+      200: UserSchema.array(),
       400: ErrorSchema,
       403: ErrorSchema,
       500: ErrorSchema,
@@ -52,15 +65,17 @@ export const userContract = contractInstance.router({
     method: 'PATCH',
     path: `${apiPrefix}/users`,
     summary: 'Patch users',
-    body: z.lazy(() => UserSchema.pick({ adminRoleIds: true, id: true }).array()),
+    body: UserSchema.pick({ adminRoleIds: true, id: true }).array(),
     description: 'Update user admin role.',
     responses: {
-      200: z.lazy(() => UserSchema.array()),
+      200: UserSchema.array(),
       400: ErrorSchema,
       403: ErrorSchema,
       500: ErrorSchema,
     },
   },
+}, {
+  baseHeaders,
 })
 
 export type LettersQuery = ClientInferRequest<typeof userContract.getMatchingUsers>['query']
