+++ packages/shared/src/contracts/zone.ts
@@ -1,11 +1,10 @@
 import type { ClientInferRequest } from '@ts-rest/core'
+import { z } from 'zod'
 import { apiPrefix, contractInstance } from '../api-client.js'
 import {
-  CreateZoneSchema,
-  DeleteZoneSchema,
-  ListZonesSchema,
-  UpdateZoneSchema,
+  ZoneSchema,
 } from '../schemas/index.js'
+import { ErrorSchema, baseHeaders } from './_utils.js'
 
 export const zoneContract = contractInstance.router({
   listZones: {
@@ -13,7 +12,13 @@ export const zoneContract = contractInstance.router({
     path: `${apiPrefix}/zones`,
     summary: 'Get zones',
     description: 'Get all zones.',
-    responses: ListZonesSchema.responses,
+    responses: {
+      200: ZoneSchema.array(),
+      400: ErrorSchema,
+      403: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   createZone: {
@@ -22,8 +27,16 @@ export const zoneContract = contractInstance.router({
     contentType: 'application/json',
     summary: 'Create zone',
     description: 'Create new zone.',
-    body: CreateZoneSchema.body,
-    responses: CreateZoneSchema.responses,
+    body: ZoneSchema
+      .omit({ id: true })
+      .extend({ clusterIds: z.string().uuid().array().optional() }),
+    responses: {
+      201: ZoneSchema,
+      400: ErrorSchema,
+      403: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   updateZone: {
@@ -31,9 +44,18 @@ export const zoneContract = contractInstance.router({
     path: `${apiPrefix}/zones/:zoneId`,
     summary: 'Update zone',
     description: 'Update a zone by its ID.',
-    pathParams: UpdateZoneSchema.params,
-    body: UpdateZoneSchema.body,
-    responses: UpdateZoneSchema.responses,
+    pathParams: z.object({
+      zoneId: z.string()
+        .uuid(),
+    }),
+    body: ZoneSchema.omit({ id: true, slug: true }),
+    responses: {
+      201: ZoneSchema,
+      400: ErrorSchema,
+      403: ErrorSchema,
+      404: ErrorSchema,
+      500: ErrorSchema,
+    },
   },
 
   deleteZone: {
@@ -41,12 +63,19 @@ export const zoneContract = contractInstance.router({
     path: `${apiPrefix}/zones/:zoneId`,
     summary: 'Delete zone',
     description: 'Delete a zone by its ID.',
-    pathParams: DeleteZoneSchema.params,
     body: null,
-    responses: DeleteZoneSchema.responses,
+    pathParams: z.object({
+      zoneId: z.string()
+        .uuid(),
+    }),
+    responses: {
+      204: null,
+      500: ErrorSchema,
+    },
   },
+}, {
+  baseHeaders,
 })
 
 export type CreateZoneBody = ClientInferRequest<typeof zoneContract.createZone>['body']
-
 export type UpdateZoneBody = ClientInferRequest<typeof zoneContract.updateZone>['body']
