+++ packages/shared/src/index.ts
@@ -1,4 +1,4 @@
 export * from './api-client.js'
-export * from './contracts/index.js'
 export * from './schemas/index.js'
+export * from './contracts/index.js'
 export * from './utils/index.js'
