+++ packages/shared/src/schemas/_utils.ts
@@ -1,17 +1,16 @@
 import { z } from 'zod'
 
-export const ErrorSchema = z.object({
-  message: z.string()
-    .optional(),
-  error: z.unknown().optional(),
-  stack: z.unknown().optional(),
-})
+export const CoerceBooleanSchema = z.boolean()
+  .or(z.enum(['true', 'false'])
+    .transform(value => value === 'true'))
 
 export const dateToString = z.string().or(z.date().transform(date => date.toISOString()))
+
 export const AtDatesToStringExtend = {
   updatedAt: dateToString,
   createdAt: dateToString,
 }
+export const permissionLevelSchema = z.coerce.string()
 
 export const UuidOrCsvUuidSchema = z.string()
   .refine((value) => {
