+++ packages/shared/src/schemas/cluster.ts
@@ -1,9 +1,4 @@
 import { z } from 'zod'
-import { EnvironmentSchema } from './environment.js'
-import { OrganizationSchema } from './organization.js'
-// import { ProjectSchemaV2 } from './project.js'
-import { UserSchema } from './user.js'
-import { ErrorSchema } from './utils.js'
 
 export const ClusterPrivacySchema = z.enum(['public', 'dedicated'])
 
@@ -65,88 +60,3 @@ export type ClusterDetails = Zod.infer<typeof ClusterDetailsSchema>
 export type Kubeconfig = Zod.infer<typeof KubeconfigSchema>
 
 export type CleanedCluster = Zod.infer<typeof CleanedClusterSchema>
-
-export const ClusterParams = z.object({
-  clusterId: CleanedClusterSchema.shape.id,
-})
-
-export const CreateClusterSchema = {
-  body: ClusterDetailsSchema.omit({ id: true }),
-  responses: {
-    201: ClusterDetailsSchema,
-    400: ErrorSchema,
-    401: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-export type CreateClusterBody = Zod.infer<typeof CreateClusterSchema.body>
-
-export const GetClustersSchema = {
-  responses: {
-    200: z.array(CleanedClusterSchema),
-    401: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-
-export const GetClusterAssociatedEnvironmentsSchema = {
-  params: ClusterParams,
-  responses: {
-    200: z.array(z.object({
-      organization: OrganizationSchema.shape.name,
-      // TODO: Remettre `ProjectSchemaV2.shape.name` mais attention aux projets non compatibles
-      project: z.string(),
-      name: EnvironmentSchema.shape.name,
-      owner: UserSchema.shape.email.optional(),
-    })),
-    401: ErrorSchema,
-    404: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-
-export const GetClusterDetailsSchema = {
-  params: ClusterParams,
-  responses: {
-    200: ClusterDetailsSchema,
-    401: ErrorSchema,
-    404: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-
-export const GetClusterKubeconfigSchema = {
-  params: ClusterParams,
-  responses: {
-    200: KubeconfigSchema,
-    401: ErrorSchema,
-    404: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-
-export const UpdateClusterSchema = {
-  params: ClusterParams,
-  body: ClusterDetailsSchema.omit({ id: true, label: true }).partial(),
-  responses: {
-    200: ClusterDetailsSchema,
-    400: ErrorSchema,
-    401: ErrorSchema,
-    404: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-
-export type UpdateClusterBody = Zod.infer<typeof UpdateClusterSchema.body>
-
-export const DeleteClusterSchema = {
-  params: ClusterParams,
-  responses: {
-    204: z.string()
-      .nullable(),
-    400: ErrorSchema,
-    401: ErrorSchema,
-    404: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
