+++ packages/shared/src/schemas/services.ts
@@ -1,7 +1,5 @@
 import { z } from 'zod'
-import { ErrorSchema } from './utils.js'
-import { pluginUpdateBody, pluginsPopulatedManifests } from './config.js'
-import { ProjectParams } from './project.js'
+import { pluginsPopulatedManifests } from './config.js'
 
 export const serviceUrl = z.object({
   to: z.string(),
@@ -22,28 +20,6 @@ export const ServiceSchema = z.object({
 })
 
 export const permissionTarget = z.enum(['user', 'admin']).default('user')
+
 export type PermissionTarget = Zod.infer<typeof permissionTarget>
-
-export const GetProjectServicesSchema = {
-  query: z.object({ permissionTarget }),
-  params: ProjectParams,
-  responses: {
-    200: ServiceSchema.array(),
-    400: ErrorSchema,
-    401: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-
-export const UpdateProjectServicesSchema = {
-  params: ProjectParams,
-  body: pluginUpdateBody,
-  responses: {
-    204: null,
-    400: ErrorSchema,
-    401: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-
 export type ProjectService = Zod.infer<typeof ServiceSchema>
