+++ packages/shared/src/schemas/zone.ts
@@ -1,5 +1,4 @@
 import { z } from 'zod'
-import { ErrorSchema } from './utils.js'
 
 export const ZoneSchema = z.object({
   id: z.string()
@@ -19,52 +18,3 @@ export const ZoneSchema = z.object({
 })
 
 export type Zone = Zod.infer<typeof ZoneSchema>
-
-export const ListZonesSchema = {
-  responses: {
-    200: z.array(ZoneSchema),
-    400: ErrorSchema,
-    403: ErrorSchema,
-    404: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-
-export const CreateZoneSchema = {
-  body: ZoneSchema
-    .omit({ id: true })
-    .extend({ clusterIds: z.string().uuid().array().optional() }),
-  responses: {
-    201: ZoneSchema,
-    400: ErrorSchema,
-    403: ErrorSchema,
-    404: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-
-export const UpdateZoneSchema = {
-  params: z.object({
-    zoneId: z.string()
-      .uuid(),
-  }),
-  body: ZoneSchema.omit({ id: true, slug: true }),
-  responses: {
-    201: ZoneSchema,
-    400: ErrorSchema,
-    403: ErrorSchema,
-    404: ErrorSchema,
-    500: ErrorSchema,
-  },
-}
-
-export const DeleteZoneSchema = {
-  params: z.object({
-    zoneId: z.string()
-      .uuid(),
-  }),
-  responses: {
-    204: null,
-    500: ErrorSchema,
-  },
-}
